+++ app/(protected)/dashboard/DashboardClient.tsx
