+++ app/api/auth/[...auth0]/route.ts
