+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import './globals.css'
-import Auth0Provider from '@/context/AuthContext';
-import LoadingProvider from '@/context/LoadingProvider';
-import ToastProvider from '@/context/ToastProvider';
+import Auth0Provider from '../context/AuthContext';
+import LoadingProvider from '../context/LoadingProvider';
+import ToastProvider from '../context/ToastProvider';
 export default function RootLayout({
   children,
 }: {
